Restructuring Code, SUCCESSFUL!
I successfully reduced the RAM & CPU USAGE By using the buffering and lazy loading. Unfortunately, I need to restructure all the code and it took me half a day. I still need to verify the action recognition model and code's ability to recognize actions. For the mean time, you can explore on how to adjust the video with slider and visualizing the existence of the student during the session. With this, you should address the vanishing incidents or the inconsistencies before frames. Also, make flowchart for easy visualization.